[update] project構成の変更
