+++ main.py
@@ -2,7 +2,7 @@ from dotenv import load_dotenv
 
 load_dotenv()
 
-from project.graph import get_graph  # noqa: E402
+from project.graph.graph import get_graph  # noqa: E402
 
 graph = get_graph()
 app = graph.compile()
