--- project/agent/__init__.py
@@ -1,7 +1,0 @@
-# ruff: noqa: F401
-from project.agent.agents.agents import agents
-from project.agent.coordinator.coordinator import coordinator
-from project.agent.evaluator.evaluator import agent_evaluator
-from project.agent.generator.generator import agent_generator
-from project.agent.optimizer.optimizer import agent_optimizer
-from project.agent.orchestrator.orchestrator import orchestrator
